ReCommited
I had a problem with the score and the player death, was reaching over 100 points before player health goes to 0, this is because the winning scene hasn't been added yet. Now I had arranged reduced the time of the bullets in order to make the score go up slower, now the problem is arranged.